Working on ensuring that screening only happens 1x
Currently, entities are able to return to screening
multiple times.  1_6 is an example in the test set.
We have an issue with our waiting list entity,
in that we don't prune the wl after an entity
receives a service.  Perhaps a different
metaphor is in order?  That, or we actively
prune an entity from waiting lists everytime
it gets assigned a service...