Fleshed out service process sim
We have the ability to express,
via spork.util.sampling functions,
complex branching and stochastic
process behavior parameterized
by various empirical distributions.

This ties in with the simple
rule specification and graph-based
process flow defined in project
tables.

Users should be able to express
a pretty sophisticated processing
scheme based on the graph relations,
including stochastic variables for
the number of children selected,
as well as the service implications
of individual nodes.