Established stacked service plans
Defined operations and behavior to
account for the plan-stack, which
may have deferred service plans.

Also added aux functions to
bypass some of the default
service-network->provider
entity creation, setting
the providers' services to be
the provider name.

This is a hack, implying we're
working with a different
semantic meaning in the service
network, but it's built on top
of the underlying services
architecture, so we should
be okay.  legacy support still
works.