Migrated to project-base xl data
Dealing with some wierdness regarding
capacities and excel processing
and table schema coercion.

I'm trying to use tabdelimited
as an intermediate format to
coerce via schema, but this is
problematic.  Excel values
are 2.0 vs. 2 for a long,
since everything's a double.

So, when parsing, the Long
parse fails, for :long?,
with a fall-back to min-value.

This defaulted to infinite
capacities.

Another problem with doubles was
the hashing.  We had dupes due
to process times being in floating
point, needlessly like above.

Solution is to coerce them
to longs and ensure no blanks
in input.

Maybe inject a schema/conformance
layer onto spork.util.excel...