Fixed processing errors from input data, verifying
Some of the constrained processing pipelines,
namely for families, aren't matching up with
the expected processing times.  Families are
able to process every 10 minutes, which not
supposed to be possible.  It should be
~35 to 55 minutes per family.  So, something
in the processing pipeline is jacked up and
allowing us to violate constraints (methinks).
Fast-track looks good.

Next step is to isolate standard and fast-track.

In general, the quantities look solid, like
we're not busting and flow constraints.

However, processing times for family
services are definitely way too fast,
we're truncating the family-services
to 1 minute vs. 5 like they're supposed
to be!